[DEV-167] Typed Node Clients (#293)
* Implemented node client types for DFI/BTC

* [DEV-167] Fixed client generation, added TODO

* [DEV-167] Check for deposit, added thunder request
