+++ src/payment/models/staking-ref-reward/staking-ref-reward.service.ts
@@ -11,11 +11,11 @@ import { Staking } from '../staking/staking.entity';
 import { ConversionService } from 'src/shared/services/conversion.service';
 import { NodeService, NodeType } from 'src/ain/node/node.service';
 import { ExchangeUtilityService } from '../exchange/exchange-utility.service';
-import { NodeClient } from 'src/ain/node/node-client';
+import { DeFiClient } from 'src/ain/node/defi-client';
 
 @Injectable()
 export class StakingRefRewardService {
-  private client: NodeClient;
+  private client: DeFiClient;
 
   constructor(
     nodeService: NodeService,
