+++ thunder-tests/thunderCollection.json
@@ -151,6 +151,13 @@
         "containerId": "",
         "created": "2022-06-21T21:40:10.043Z",
         "sortNum": 210000
+      },
+      {
+        "_id": "0dfedb35-f836-48a1-b010-04736f083788",
+        "name": "Crypto Route",
+        "containerId": "",
+        "created": "2022-07-08T16:28:33.047Z",
+        "sortNum": 220000
       }
     ],
     "settings": {
