+++ thunder-tests/thunderclient.json
@@ -1236,5 +1236,38 @@
       ]
     },
     "tests": []
+  },
+  {
+    "_id": "62b4036a-686f-4fd6-bb82-a712e32ca1b7",
+    "colId": "62eefab7-84cd-496b-8f93-253c15d4f841",
+    "containerId": "0dfedb35-f836-48a1-b010-04736f083788",
+    "name": "Get My Crypto Routes",
+    "url": "{{url}}/v1/cryptoRoute",
+    "method": "GET",
+    "sortNum": 410000,
+    "created": "2022-07-08T16:28:47.111Z",
+    "modified": "2022-07-08T16:29:06.302Z",
+    "headers": [],
+    "params": [],
+    "tests": []
+  },
+  {
+    "_id": "a6f50f94-33d5-4626-bcf4-abc196db233c",
+    "colId": "62eefab7-84cd-496b-8f93-253c15d4f841",
+    "containerId": "0dfedb35-f836-48a1-b010-04736f083788",
+    "name": "Create Crypto Route",
+    "url": "{{url}}/v1/cryptoRoute",
+    "method": "POST",
+    "sortNum": 420000,
+    "created": "2022-07-08T16:29:53.792Z",
+    "modified": "2022-07-08T16:49:16.378Z",
+    "headers": [],
+    "params": [],
+    "body": {
+      "type": "json",
+      "raw": "{\n  \"buyType\": \"Wallet\",\n  \"blockchain\": \"Bitcoin\",\n  \"asset\": {\n      \"id\": 2\n  }\n}",
+      "form": []
+    },
+    "tests": []
   }
 ]
